third_party/WebKit/LayoutTests/fast/dom/call-a-constructor-as-a-function-expected.txt

This tests if TypeError is thrown or not when we call a constructor as a normal function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ArrayBuffer() threw exception TypeError: Constructor ArrayBuffer requires 'new'.
PASS AudioContext() threw exception TypeError: Failed to construct 'AudioContext': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS FormData() threw exception TypeError: Failed to construct 'FormData': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS DataView() threw exception TypeError: Constructor DataView requires 'new'.
PASS EventSource() threw exception TypeError: Failed to construct 'EventSource': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS FileReader() threw exception TypeError: Failed to construct 'FileReader': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS Float32Array() threw exception TypeError: Constructor Float32Array requires 'new'.
PASS Float64Array() threw exception TypeError: Constructor Float64Array requires 'new'.
PASS Audio() threw exception TypeError: Failed to construct 'Audio': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS Image() threw exception TypeError: Failed to construct 'Image': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS Option() threw exception TypeError: Failed to construct 'Option': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS Int16Array() threw exception TypeError: Constructor Int16Array requires 'new'.
PASS Int32Array() threw exception TypeError: Constructor Int32Array requires 'new'.
PASS Int8Array() threw exception TypeError: Constructor Int8Array requires 'new'.
PASS MessageChannel() threw exception TypeError: Failed to construct 'MessageChannel': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS SharedWorker() threw exception TypeError: Failed to construct 'SharedWorker': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS Uint16Array() threw exception TypeError: Constructor Uint16Array requires 'new'.
PASS Uint32Array() threw exception TypeError: Constructor Uint32Array requires 'new'.
PASS Uint8Array() threw exception TypeError: Constructor Uint8Array requires 'new'.
PASS WebKitCSSMatrix() threw exception TypeError: Failed to construct 'DOMMatrix': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS WebSocket() threw exception TypeError: Failed to construct 'WebSocket': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS Worker() threw exception TypeError: Failed to construct 'Worker': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS XMLHttpRequest() threw exception TypeError: Failed to construct 'XMLHttpRequest': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS XSLTProcessor() threw exception TypeError: Failed to construct 'XSLTProcessor': Please use the 'new' operator, this DOM object constructor cannot be called as a function..
PASS successfullyParsed is true

TEST COMPLETE
